requirements.txt
Flask==3.0.0
Flask-JWT-Extended==4.5.5
Flask-Cors==4.0.1
psycopg2-binary==2.9.9
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
xgboost==2.1.0
joblib==1.4.2
python-dotenv==1.0.1
requests==2.31.0
geopy==2.4.1